- Added patches.suse/perf-{0001-0279}*.patch (FATE#311392, BNC#685313) - patches.xen/kexec-move-segment-code-i386.patch: Refresh. - patches.xen/xen3-auto-arch-i386.diff: Refresh. - patches.xen/xen3-auto-arch-x86.diff: Refresh. - patches.xen/xen3-auto-arch-x86_64.diff: Refresh. - patches.xen/xen3-patch-2.6.25: Refresh. - patches.xen/xen3-patch-2.6.27: Refresh. - patches.xen/xen3-patch-2.6.28: Refresh. - patches.xen/xen3-patch-2.6.30: Refresh. - patches.xen/xen3-patch-2.6.31: Refresh. - patches.xen/xen3-patch-2.6.32.11-12: Refresh. - Xen sequences but does not build.
suse-commit: 6a1603cfe787ed3e3ccb69335fe62f3776adfe4f

+++ kernel/pid.c
@@ -516,9 +516,9 @@ void __init pidmap_init(void)
 {
 	/* bump default and minimum pid_max based on number of cpus */
 	pid_max = min(pid_max_max, max(pid_max,
-				PIDS_PER_CPU_DEFAULT * num_possible_cpus()));
+				(int)(PIDS_PER_CPU_DEFAULT * num_possible_cpus())));
 	pid_max_min = max(pid_max_min,
-				PIDS_PER_CPU_MIN * num_possible_cpus());
+			  (int)(PIDS_PER_CPU_MIN * num_possible_cpus()));
 	pr_info("pid_max: default: %u minimum: %u\n", pid_max, pid_max_min);
 
 	init_pid_ns.pidmap[0].page = kzalloc(PAGE_SIZE, GFP_KERNEL);
